requirements.txt
plaintext
cassandra-driver
openai
pytest
